додав роутер для Кп тепер можна витягувати за name address rating
