+++ Backend/src/project/repository/utility_company_rep.py
@@ -1,0 +1,6 @@
+from sqlalchemy.orm import Session
+from ..db import models
+
+def all(db:Session):
+  companies = db.query(models.UtilityCompany).all()
+  return companies
